EL porcentaje para acabar la tarea ahora es un entero, ya que la barra de progreso funciona con enteros

+++ HttpImageDownloader/src/es/ucm/si/dneb/service/gestionTareas/ServicioGestionTareas.java
@@ -21,7 +21,7 @@ public interface ServicioGestionTareas {
 	
 	public void pararTarea(long tareaId);
 	
-	public double obtenerPorcentajeCompletado(long tareaId);
+	public int obtenerPorcentajeCompletado(long tareaId);
 		
 	public List<Survey> getAllSurveys();
